feat: scanNullableTime関数とparseTimeString関数を新規作成し、NULL許容のtime.Timeを返す機能を追加
