+++ internal/infra/repository/song_repository_impl.go
@@ -155,64 +155,6 @@ func (r *songRepository) GetLatestUpdatedAtExcludingWorldsend(ctx context.Contex
 	return scanNullableTime(ctx, exec, query)
 }
 
-func scanNullableTime(ctx context.Context, exec repository.Executor, query string, args ...any) (*time.Time, error) {
-	var raw any
-	if err := exec.QueryRowxContext(ctx, query, args...).Scan(&raw); err != nil {
-		return nil, err
-	}
-	if raw == nil {
-		return nil, nil
-	}
-
-	switch value := raw.(type) {
-	case time.Time:
-		updatedAt := value
-		return &updatedAt, nil
-	case []byte:
-		return parseTimeString(string(value))
-	case string:
-		return parseTimeString(value)
-	default:
-		return nil, fmt.Errorf("unsupported updated_at type: %T", raw)
-	}
-}
-
-func parseTimeString(value string) (*time.Time, error) {
-	return parseTimeStringInLocation(value, time.Local)
-}
-
-func parseTimeStringInLocation(value string, location *time.Location) (*time.Time, error) {
-	trimmed := strings.TrimSpace(value)
-	if trimmed == "" {
-		return nil, nil
-	}
-
-	zonedLayouts := []string{
-		time.RFC3339Nano,
-		"2006-01-02 15:04:05.999999999Z07:00",
-		"2006-01-02 15:04:05Z07:00",
-	}
-	for _, layout := range zonedLayouts {
-		parsed, err := time.Parse(layout, trimmed)
-		if err == nil {
-			return &parsed, nil
-		}
-	}
-
-	localLayouts := []string{
-		"2006-01-02 15:04:05.999999999",
-		"2006-01-02 15:04:05",
-	}
-	for _, layout := range localLayouts {
-		parsed, err := time.ParseInLocation(layout, trimmed, location)
-		if err == nil {
-			return &parsed, nil
-		}
-	}
-
-	return nil, fmt.Errorf("failed to parse updated_at: %s", trimmed)
-}
-
 func (r *songRepository) toSongEntity(row *songRow) *entity.Song {
 	song := entity.NewSong()
 	song.ID = row.ID
